moved shared classes to a separate utils file
