+++ baseclasses/BaseSolver.py
@@ -1,74 +1,15 @@
-from pprint import pprint as pp
-
 """
 BaseSolver
 
 Holds a basic Python Analysis Classes (base and inherited).
-
-Copyright (c) 2017 by Dr. Charles A. Mader
-All rights reserved. Not to be used for commercial purposes.
-Revision: 1.0   $Date: 01/06/2017 15:00$
-
-
-Developers:
------------
-- Dr. Charles A. Mader (CAM)
-
-History
--------
-    v. 1.0    - Initial Class Creation (CAM, 2013)
-    v. 2.0    - Major update to options implementation (CAM,2017)
 """
-
-__version__ = "$Revision: $"
-
-
-# =============================================================================
-# Standard Python modules
-# =============================================================================
-
-# =============================================================================
-# Misc Definitions
-# =============================================================================
-
-
-class CaseInsensitiveDict(dict):
-    def __setitem__(self, key, value):
-        super(CaseInsensitiveDict, self).__setitem__(key.lower(), value)
-
-    def __getitem__(self, key):
-        return super(CaseInsensitiveDict, self).__getitem__(key.lower())
-
-    def __contains__(self, key):
-        return super(CaseInsensitiveDict, self).__contains__(key.lower())
-
-
-class Error(Exception):
-    """
-    Format the error message in a box to make it clear this
-    was a explicitly raised exception.
-    """
-
-    def __init__(self, message):
-        msg = "\n+" + "-" * 78 + "+" + "\n" + "| BaseSolver Error: "
-        i = 19
-        for word in message.split():
-            if len(word) + i + 1 > 78:  # Finish line and start new one
-                msg += " " * (78 - i) + "|\n| " + word + " "
-                i = 1 + len(word) + 1
-            else:
-                msg += word + " "
-                i += len(word) + 1
-        msg += " " * (78 - i) + "|\n" + "+" + "-" * 78 + "+" + "\n"
-        print(msg)
-        Exception.__init__(self)
-
+from pprint import pprint as pp
+from .utils import CaseInsensitiveDict, Error
 
 # =============================================================================
 # BaseSolver Class
 # =============================================================================
 class BaseSolver(object):
-
     """
     Abstract Class for a basic Solver Object
     """
@@ -76,8 +17,6 @@ class BaseSolver(object):
     def __init__(self, name, category={}, def_options={}, **kwargs):
         """
         Solver Class Initialization
-
-        Documentation last updated:
         """
 
         self.name = name
@@ -105,8 +44,6 @@ class BaseSolver(object):
     def __call__(self, *args, **kwargs):
         """
         Run Analyzer (Calling Routine)
-
-        Documentation last updated:
         """
 
         # Checks
