+++ baseclasses/pyAeroStruct_problem.py
@@ -1,40 +1,12 @@
 """
 pyAeroStruct_problem
-
-Developers:
------------
-- Dr. Gaetan K. W. Kenway (GKWK)
-
-History
--------
-    v. 0.1    - Complete overall of AeroProblem (GKWK, 2014)
 """
 # ======================================================================
 # Imports
 # ======================================================================
 from .pyAero_problem import AeroProblem
 from .pyStruct_problem import StructProblem
-
-
-class Error(Exception):
-    """
-    Format the error message in a box to make it clear this
-    was a expliclty raised exception.
-    """
-
-    def __init__(self, message):
-        msg = "\n+" + "-" * 78 + "+" + "\n" + "| AeroStructProblem Error: "
-        i = 26
-        for word in message.split():
-            if len(word) + i + 1 > 78:  # Finish line and start new one
-                msg += " " * (78 - i) + "|\n| " + word + " "
-                i = 1 + len(word) + 1
-            else:
-                msg += word + " "
-                i += len(word) + 1
-        msg += " " * (78 - i) + "|\n" + "+" + "-" * 78 + "+" + "\n"
-        print(msg)
-        Exception.__init__(self)
+from .utils import Error
 
 
 class AeroStructProblem(object):
