+++ baseclasses/pyAero_problem.py
@@ -1,13 +1,5 @@
 """
 pyAero_problem
-
-Developers:
------------
-- Dr. Gaetan K. W. Kenway (GKWK)
-
-History
--------
-    v. 0.1    - Complete overall of AeroProblem (GKWK, 2014)
 """
 
 # =============================================================================
@@ -17,35 +9,7 @@ import numpy
 import warnings
 from .ICAOAtmosphere import ICAOAtmosphere
 from .FluidProperties import FluidProperties
-
-
-class CaseInsensitiveDict(dict):
-    def __setitem__(self, key, value):
-        super(CaseInsensitiveDict, self).__setitem__(key.lower(), value)
-
-    def __getitem__(self, key):
-        return super(CaseInsensitiveDict, self).__getitem__(key.lower())
-
-
-class Error(Exception):
-    """
-    Format the error message in a box to make it clear this
-    was a expliclty raised exception.
-    """
-
-    def __init__(self, message):
-        msg = "\n+" + "-" * 78 + "+" + "\n" + "| AeroProblem Error: "
-        i = 20
-        for word in message.split():
-            if len(word) + i + 1 > 78:  # Finish line and start new one
-                msg += " " * (78 - i) + "|\n| " + word + " "
-                i = 1 + len(word) + 1
-            else:
-                msg += word + " "
-                i += len(word) + 1
-        msg += " " * (78 - i) + "|\n" + "+" + "-" * 78 + "+" + "\n"
-        print(msg)
-        Exception.__init__(self)
+from .utils import CaseInsensitiveDict, Error
 
 
 class AeroProblem(FluidProperties):
@@ -240,7 +204,7 @@ areaRef=0.772893541, chordRef=0.64607, xRef=0.0, zRef=0.0, alpha=3.06, T=255.56)
     def __init__(self, name, **kwargs):
 
         # Set basic fluid properties
-        FluidProperties.__init__(self, **kwargs)
+        super().__init__(self, **kwargs)
 
         # Always have to have the name
         self.name = name
