+++ baseclasses/pyFieldPerformance_problem.py
@@ -1,14 +1,5 @@
 """
 pyFieldPerformance_problem
-
-Developers:
------------
-- Dr. Charles A.(Sandy) Mader (CAM)
-- Nicolas Bons   (NB)
-
-History
--------
-    v. 0.1    - Initial Implementation of FieldPerformance Problem
 """
 
 # =============================================================================
@@ -17,35 +8,7 @@ History
 import warnings
 from .ICAOAtmosphere import ICAOAtmosphere
 from .FluidProperties import FluidProperties
-
-
-class CaseInsensitiveDict(dict):
-    def __setitem__(self, key, value):
-        super(CaseInsensitiveDict, self).__setitem__(key.lower(), value)
-
-    def __getitem__(self, key):
-        return super(CaseInsensitiveDict, self).__getitem__(key.lower())
-
-
-class Error(Exception):
-    """
-    Format the error message in a box to make it clear this
-    was a expliclty raised exception.
-    """
-
-    def __init__(self, message):
-        msg = "\n+" + "-" * 78 + "+" + "\n" + "| FieldPerformanceProblem Error: "
-        i = 20
-        for word in message.split():
-            if len(word) + i + 1 > 78:  # Finish line and start new one
-                msg += " " * (78 - i) + "|\n| " + word + " "
-                i = 1 + len(word) + 1
-            else:
-                msg += word + " "
-                i += len(word) + 1
-        msg += " " * (78 - i) + "|\n" + "+" + "-" * 78 + "+" + "\n"
-        print(msg)
-        Exception.__init__(self)
+from .utils import CaseInsensitiveDict, Error
 
 
 class FieldPerformanceProblem(object):
