+++ baseclasses/pyLG_problem.py
@@ -8,46 +8,9 @@ computation. Nothing in this computation can be a design variable right now beca
 the beam model setup in TACS has no provision for changing the load as a design variable.
 We will assume a fixed aircraft mass and LG characteristics. The output load can change as a
 function of the LG geometry if necessary.
-
-Copyright (c) 2019 by Dr. Charles A. Mader
-All rights reserved. Not to be used for commercial purposes.
-Revision: 1.0   $Date: 12/02/2019 21:00$
-
-
-Developers:
------------
-- Dr. Charles A. Mader (CM)
-
-History
--------
-    v. 1.0 - Initial Class Creation (CM, 2019)
-
 """
 
 import numpy
-
-# from pygeo import geo_utils
-
-
-class Error(Exception):
-    """
-    Format the error message in a box to make it clear this
-    was a expliclty raised exception.
-    """
-
-    def __init__(self, message):
-        msg = "\n+" + "-" * 78 + "+" + "\n" + "| LGProblem Error: "
-        i = 23
-        for word in message.split():
-            if len(word) + i + 1 > 78:  # Finish line and start new one
-                msg += " " * (78 - i) + "|\n| " + word + " "
-                i = 1 + len(word) + 1
-            else:
-                msg += word + " "
-                i += len(word) + 1
-        msg += " " * (78 - i) + "|\n" + "+" + "-" * 78 + "+" + "\n"
-        print(msg)
-        Exception.__init__(self)
 
 
 class LGProblem(object):
