+++ baseclasses/pyStruct_problem.py
@@ -1,36 +1,12 @@
 """
 pyStrut_problem
-
-Developers:
------------
-- Dr. Gaetan K. W. Kenway (GKWK)
 """
 
 # =============================================================================
 # Imports
 # =============================================================================
 import warnings
-
-
-class Error(Exception):
-    """
-    Format the error message in a box to make it clear this
-    was a expliclty raised exception.
-    """
-
-    def __init__(self, message):
-        msg = "\n+" + "-" * 78 + "+" + "\n" + "| StructProblem Error: "
-        i = 21
-        for word in message.split():
-            if len(word) + i + 1 > 78:  # Finish line and start new one
-                msg += " " * (78 - i) + "|\n| " + word + " "
-                i = 1 + len(word) + 1
-            else:
-                msg += word + " "
-                i += len(word) + 1
-        msg += " " * (78 - i) + "|\n" + "+" + "-" * 78 + "+" + "\n"
-        print(msg)
-        Exception.__init__(self)
+from .utils import Error
 
 
 class StructProblem(object):
